renamed Helper classes namespace
