+++ src/Downloader.Test/GlobalUsings.cs
@@ -16,4 +16,3 @@ global using System.Reflection;
 global using Microsoft.Extensions.Logging;
 global using Newtonsoft.Json;
 global using Downloader.Test.Helper;
-global using Downloader.Extensions.Helpers;
