+++ src/Downloader/DownloadService.cs
@@ -1,4 +1,4 @@
-using Downloader.Extensions.Helpers;
+using Downloader.Extensions;
 using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
